requirements.txt
requests
beautifulsoup4
google-cloud-storage